backend/requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiogram_dialog==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
attrs==25.3.0
bcrypt==5.0.0
billiard==4.2.2
black==25.9.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.1
Django==5.2.6
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fastapi==0.118.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==6.1.0
Jinja2==3.1.6
kombu==5.5.4
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.6.4
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
psycopg2-binary==2.9.10
pwdlib==0.2.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-pika==0.9.14
pytokens==0.1.10
redis==6.4.0
requests==2.32.5
rest-framework-simplejwt==0.0.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.20.1
